requirements.txt
streamlit== 1.29.0
pypdf2==
langchain==
python-dotenv== 
faiss-cpu== 
openai==
huggingface_hub==
InstructorEmbedding==
sentence_transformers==
pypdf2
gtts
playsound
pyaudio
elevenlabs
